[z0362] .debug setitemvalue/getitemvalues commands (removed setitemflag).
Also backporting some code cleanups from original commit.

(based on master commit 55eefba)

(based on commit b939551)

+++ src/game/Player.cpp
@@ -12794,7 +12794,7 @@ void Player::SendQuestTimerFailed( uint32 quest_id )
     if( quest_id )
     {
         WorldPacket data( SMSG_QUESTUPDATE_FAILEDTIMER, 4 );
-        data << quest_id;
+        data << uint32(quest_id);
         GetSession()->SendPacket( &data );
         sLog.outDebug("WORLD: Sent SMSG_QUESTUPDATE_FAILEDTIMER");
     }
